requirements.txt
# Only for testing
pytest

# Optional
pyproj

# Needed for library to work
requests
pillow
